# API-SPECIFICATION.yml
openapi: 3.0.0
paths:
...
  /api/v1/auth/register:
    post:
      operationId: AuthController_register
      summary: Register a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDto"
      responses:
        201:
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponseDto"
        400:
          description: Validation error
        409:
          description: User already exists
      tags:
        - Authentication
  /api/v1/auth/login:
    post:
      operationId: AuthController_login
      summary: Login user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponseDto"
        401:
          description: Invalid credentials
      tags:
        - Authentication
  /api/v1/auth/logout:
    post:
      operationId: AuthController_logout
      summary: Logout user (invalidate session)
      parameters: []
      responses:
        204:
          description: Logout successful
        401:
          description: Unauthorized
      tags:
        - Authentication
      security:
        -
          JWT-auth: []
  /api/v1/auth/me:
    get:
      operationId: AuthController_getCurrentUser
      summary: Get current user profile
      parameters: []
      responses:
        200:
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        401:
          description: Unauthorized
      tags:
        - Authentication
      security:
        -
          JWT-auth: []
...
  /api/v1/orders:
    post:
      operationId: OrderController_createOrder
      summary: Create a new order
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderDto"
      responses:
        201:
          description: Order created successfully
        400:
          description: Invalid request or restaurant closed
        404:
          description: Restaurant or menu items not found
      tags:
        - Orders
      security:
        -
          JWT-auth: []
    get:
      operationId: OrderController_getOrderHistory
      summary: Get order history
      parameters:
        -
          name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            example: 1
            type: number
        -
          name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 10
            example: 10
            type: number
        -
          name: status
          required: false
          in: query
          schema:
            enum:
              - PENDING
              - DELIVERED
              - CANCELLED
            type: string
      responses:
        200:
          description: Order history retrieved
      tags:
        - Orders
      security:
        -
          JWT-auth: []
  /api/v1/orders/{id}:
    get:
      operationId: OrderController_getOrderById
      summary: Get order details by ID (includes payment info)
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: Order details retrieved
        404:
          description: Order not found
      tags:
        - Orders
      security:
        -
          JWT-auth: []
  /api/v1/orders/{id}/cancel:
    patch:
      operationId: OrderController_cancelOrder
      summary: Cancel an order (within 1 minute of creation)
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderDto"
      responses:
        200:
          description: Order cancelled successfully
        400:
          description: Cancellation window expired or invalid status
        404:
          description: Order not found
      tags:
        - Orders
      security:
        -
          JWT-auth: []
...
  /api/v1/driver/profile:
    post:
      operationId: DriverController_createProfile
      summary: Create driver profile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverDto"
      responses:
        201:
          description: Driver profile created successfully
        400:
          description: Driver profile already exists or invalid data
      tags:
        - Driver
      security:
        -
          JWT-auth: []
    get:
      operationId: DriverController_getProfile
      summary: Get driver profile
      parameters: []
      responses:
        200:
          description: Driver profile retrieved
        404:
          description: Driver profile not found
      tags:
        - Driver
      security:
        -
          JWT-auth: []
    patch:
      operationId: DriverController_updateProfile
      summary: Update driver profile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverDto"
      responses:
        200:
          description: Driver profile updated successfully
        400:
          description: Invalid data
        404:
          description: Driver profile not found
      tags:
        - Driver
      security:
        -
          JWT-auth: []
  /api/v1/driver/status:
    patch:
      operationId: DriverController_updateStatus
      summary: Toggle driver online/offline status
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverStatusDto"
      responses:
        200:
          description: Driver status updated successfully
        403:
          description: Driver not approved
        404:
          description: Driver profile not found
      tags:
        - Driver
      security:
        -
          JWT-auth: []
  /api/v1/driver/location:
    patch:
      operationId: DriverController_updateLocation
      summary: Update driver current location
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverLocationDto"
      responses:
        200:
          description: Driver location updated successfully
        404:
          description: Driver profile not found
      tags:
        - Driver
      security:
        -
          JWT-auth: []
    post:
      operationId: LocationController_updateLocation
      summary: Update driver GPS location
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationDto"
      responses:
        204:
          description: Location updated successfully
        400:
          description: Invalid location data
        404:
          description: Driver not found
      tags:
        - Location (Driver)
      security:
        -
          JWT-auth: []
  /api/v1/driver/earnings:
    get:
      operationId: DriverController_getEarnings
      summary: Get driver earnings and statistics
      parameters: []
      responses:
        200:
          description: Driver earnings retrieved
        404:
          description: Driver profile not found
      tags:
        - Driver
      security:
        -
          JWT-auth: []
  /api/v1/driver/location/history:
    get:
      operationId: LocationController_getHistory
      summary: Get driver location history
      parameters:
        -
          name: limit
          required: false
          in: query
          description: "Number of records to retrieve (default: 100)"
          schema:
            type: number
      responses:
        200:
          description: Location history retrieved
      tags:
        - Location (Driver)
      security:
        -
          JWT-auth: []
  /api/v1/driver/location/current:
    get:
      operationId: LocationController_getCurrentLocation
      summary: Get driver current location
      parameters: []
      responses:
        200:
          description: Current location retrieved
        404:
          description: Location not available
      tags:
        - Location (Driver)
      security:
        -
          JWT-auth: []
  /api/v1/driver/location/stats:
    get:
      operationId: LocationController_getBufferStats
      summary: Get location buffer statistics
      parameters: []
      responses:
        200:
          description: Buffer statistics retrieved
      tags:
        - Location (Driver)
      security:
        -
          JWT-auth: []
info:
  title: SwiftEats API
  description: Real-Time Food Delivery Platform API
  version: 1.0
  contact:
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          example: john.doe@example.com
        password:
          type: string
          example: Password123!
        name:
          type: string
          example: John Doe
        phone:
          type: string
          example: +919876543210
        role:
          type: string
          enum:
            - CUSTOMER
            - RESTAURANT
            - DRIVER
            - ADMIN
          example: CUSTOMER
      required:
        - email
        - password
        - name
        - phone
        - role
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum:
            - CUSTOMER
            - RESTAURANT
            - DRIVER
            - ADMIN
        isActive:
          type: boolean
        createdAt:
          format: date-time
          type: string
        lastLoginAt:
          format: date-time
          type: string
      required:
        - id
        - email
        - name
        - phone
        - role
        - isActive
        - createdAt
    LoginResponseDto:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/UserResponseDto"
        expiresIn:
          type: number
      required:
        - accessToken
        - user
        - expiresIn
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: abhinav.kumar@talentica.com
        password:
          type: string
          example: Password123!
      required:
        - email
        - password
...
    OrderItemDto:
      type: object
      properties:
        menuItemId:
          type: string
          example: uuid-of-menu-item
        quantity:
          type: number
          example: 2
          minimum: 1
        specialInstructions:
          type: string
          example: No onions please
          maxLength: 500
      required:
        - menuItemId
        - quantity
    CreateOrderDto:
      type: object
      properties:
        restaurantId:
          type: string
          example: uuid-of-restaurant
        deliveryAddressId:
          type: string
          example: uuid-of-delivery-address
        items:
          example:
            -
              menuItemId: uuid
              quantity: 2
              specialInstructions: Extra spicy
          type: array
          items:
            $ref: "#/components/schemas/OrderItemDto"
        paymentMethod:
          type: string
          enum:
            - CASH
            - CARD
            - UPI
            - WALLET
          example: CARD
        specialInstructions:
          type: string
          example: Please ring the doorbell
          maxLength: 1000
        promoCode:
          type: string
          example: WELCOME10
          maxLength: 50
      required:
        - restaurantId
        - deliveryAddressId
        - items
        - paymentMethod
    CancelOrderDto:
      type: object
      properties:
        reason:
          type: string
          example: Changed my mind
          maxLength: 500
    CreateDriverDto:
      type: object
      properties:
        licenseNumber:
          type: string
          description: Driver license number
          example: DL-1234567890123
        vehicleType:
          type: string
          description: Vehicle type
          enum:
            - BIKE
            - SCOOTER
            - CAR
          example: BIKE
        vehicleNumber:
          type: string
          description: Vehicle registration number
          example: KA-01-AB-1234
        vehicleModel:
          type: string
          description: Vehicle model name
          example: Honda Activa 6G
      required:
        - licenseNumber
        - vehicleType
        - vehicleNumber
    UpdateDriverDto:
      type: object
      properties:
        licenseNumber:
          type: string
          description: Driver license number
          example: DL-1234567890123
        vehicleType:
          type: string
          description: Vehicle type
          enum:
            - BIKE
            - SCOOTER
            - CAR
          example: BIKE
        vehicleNumber:
          type: string
          description: Vehicle registration number
          example: KA-01-AB-1234
        vehicleModel:
          type: string
          description: Vehicle model name
          example: Honda Activa 6G
    UpdateDriverStatusDto:
      type: object
      properties:
        isOnline:
          type: boolean
          description: Driver online status
          example: true
      required:
        - isOnline
    UpdateDriverLocationDto:
      type: object
      properties:
        latitude:
          type: number
          description: Latitude
          example: 12.9716
          minimum: -90
          maximum: 90
        longitude:
          type: number
          description: Longitude
          example: 77.5946
          minimum: -180
          maximum: 180
      required:
        - latitude
        - longitude
    UpdateLocationDto:
      type: object
      properties:
        latitude:
          type: number
          description: Latitude
          example: 12.9716
          minimum: -90
          maximum: 90
        longitude:
          type: number
          description: Longitude
          example: 77.5946
          minimum: -180
          maximum: 180
        accuracy:
          type: number
          description: GPS accuracy in meters
          example: 10
        heading:
          type: number
          description: Heading/bearing in degrees
          example: 180
        speed:
          type: number
          description: Speed in km/h
          example: 25
      required:
        - latitude
        - longitude